+++ zinc-compiler/src/semantic/element/value/array/mod.rs
@@ -182,6 +182,6 @@ impl Array {
 
 impl fmt::Display for Array {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "array of '{}'s", self.r#type)
+        write!(f, "<array> of '{}'s", self.r#type)
     }
 }