+++ zinc-compiler/src/semantic/element/value/integer/error.rs
@@ -19,6 +19,9 @@ pub enum Error {
     TypesMismatchDivision { first: String, second: String },
     TypesMismatchRemainder { first: String, second: String },
 
+    OperatorBitwiseShiftLeftSecondOperatorExpectedUnsigned { found: String },
+    OperatorBitwiseShiftRightSecondOperatorExpectedUnsigned { found: String },
+
     ForbiddenFieldDivision,
     ForbiddenFieldRemainder,
     ForbiddenFieldBitwise,